Remove GraphHopper#loadGraph and Princeton performance test
