--- core/src/test/resources/com/graphhopper/reader/tinyEWD.txt
@@ -1,17 +1,0 @@
-8
-15
-4 5 0.35
-5 4 0.35
-4 7 0.37
-5 7 0.28
-7 5 0.28
-5 1 0.32
-0 4 0.38
-0 2 0.26
-7 3 0.39
-1 3 0.29
-2 7 0.34
-6 2 0.40
-3 6 0.52
-6 0 0.58
-6 4 0.93
