+++ reader-osm/src/test/java/com/graphhopper/reader/osm/GraphHopperOSMTest.java
@@ -755,182 +755,6 @@ public class GraphHopperOSMTest {
     }
 
     @Test
-    public void testGetPathsDirectionEnforcement1() {
-        // Test enforce start direction
-        // Note: This Test does not pass for CH enabled    
-        instance = createSquareGraphInstance();
-
-        // Start in middle of edge 4-5 
-        GHPoint start = new GHPoint(0.0015, 0.002);
-        // End at middle of edge 2-3
-        GHPoint end = new GHPoint(0.002, 0.0005);
-
-        GHRequest req = new GHRequest().
-                setPoints(Arrays.asList(start, end)).
-                setHeadings(Arrays.asList(180., Double.NaN)).
-                setProfile("profile").
-                setPathDetails(Arrays.asList("edge_key"));
-        GHResponse response = instance.route(req);
-        assertFalse(response.getErrors().toString(), response.hasErrors());
-        assertArrayEquals(new int[]{4, 5, 8, 3, 2}, calcNodes(instance, response.getAll().get(0)));
-    }
-
-    @Test
-    public void testGetPathsDirectionEnforcement2() {
-        // Test enforce south start direction and east end direction
-        instance = createSquareGraphInstance();
-
-        // Start in middle of edge 4-5 
-        GHPoint start = new GHPoint(0.0015, 0.002);
-        // End at middle of edge 2-3
-        GHPoint end = new GHPoint(0.002, 0.0005);
-
-        GHRequest req = new GHRequest(start, end).
-                setHeadings(Arrays.asList(180.0, 90.0)).
-                setProfile("profile").
-                setPathDetails(Arrays.asList("edge_key"));
-        GHResponse response = instance.route(req);
-        assertFalse(response.hasErrors());
-        assertArrayEquals(new int[]{4, 5, 8, 1, 2, 3}, calcNodes(instance, response.getAll().get(0)));
-
-        // Test uni-directional case
-        req.setAlgorithm(DIJKSTRA);
-        response = instance.route(req);
-        assertFalse(response.getErrors().toString(), response.hasErrors());
-        assertArrayEquals(new int[]{4, 5, 8, 1, 2, 3}, calcNodes(instance, response.getAll().get(0)));
-    }
-
-    @Test
-    public void testGetPathsDirectionEnforcement3() {
-        instance = createSquareGraphInstance();
-
-        // Start in middle of edge 4-5 
-        GHPoint start = new GHPoint(0.0015, 0.002);
-        // Via Point between 8-7
-        GHPoint via = new GHPoint(0.0005, 0.001);
-        // End at middle of edge 2-3
-        GHPoint end = new GHPoint(0.002, 0.0005);
-
-        GHRequest req = new GHRequest().
-                setPoints(Arrays.asList(start, via, end)).
-                setHeadings(Arrays.asList(Double.NaN, 0., Double.NaN)).
-                setProfile("profile").
-                setPathDetails(Arrays.asList("edge_key"));
-        GHResponse response = instance.route(req);
-        assertFalse(response.hasErrors());
-        assertArrayEquals(new int[]{4, 5, 6, 7, 8, 3, 2}, calcNodes(instance, response.getAll().get(0)));
-    }
-
-    @Test
-    public void testGetPathsDirectionEnforcement4() {
-        // Test straight via routing
-        instance = createSquareGraphInstance();
-
-        // Start in middle of edge 4-5 
-        GHPoint start = new GHPoint(0.0015, 0.002);
-        // End at middle of edge 2-3
-        GHPoint end = new GHPoint(0.002, 0.0005);
-        // Via Point between 8-3
-        GHPoint via = new GHPoint(0.0015, 0.001);
-        GHRequest req = new GHRequest().
-                setPoints(Arrays.asList(start, via, end)).
-                setProfile("profile").
-                setPathDetails(Arrays.asList("edge_key"));
-        req.putHint(Routing.PASS_THROUGH, true);
-        GHResponse response = instance.route(req);
-        assertFalse(response.hasErrors());
-        assertEquals(1, response.getAll().size());
-        assertArrayEquals(new int[]{5, 4, 3, 8, 1, 2, 3}, calcNodes(instance, response.getAll().get(0)));
-    }
-
-    @Test
-    public void testGetPathsDirectionEnforcement5() {
-        // Test independence of previous enforcement for subsequent paths
-        instance = createSquareGraphInstance();
-
-        // Start in middle of edge 4-5 
-        GHPoint start = new GHPoint(0.0015, 0.002);
-        // End at middle of edge 2-3
-        GHPoint end = new GHPoint(0.002, 0.0005);
-        // First go south and than come from west to via-point at 7-6. Then go back over previously punished (11)-4 edge
-        GHPoint via = new GHPoint(0.000, 0.0015);
-        GHRequest req = new GHRequest().
-                setPoints(Arrays.asList(start, via, end)).
-                setHeadings(Arrays.asList(0., 3.14 / 2, Double.NaN)).
-                setProfile("profile").
-                setPathDetails(Arrays.asList("edge_key"));
-        req.putHint(Routing.PASS_THROUGH, true);
-        GHResponse response = instance.route(req);
-        assertFalse(response.hasErrors());
-        assertArrayEquals(new int[]{5, 4, 3, 8, 7, 6, 5, 4, 3, 2}, calcNodes(instance, response.getAll().get(0)));
-    }
-
-    @Test
-    public void testGetPathsDirectionEnforcement6() {
-        // Test if snaps at tower nodes are ignored
-        instance = createSquareGraphInstance();
-
-        // QueryPoints directly on TowerNodes 
-        GHPoint start = new GHPoint(0, 0);
-        GHPoint via = new GHPoint(0.002, 0.000);
-        GHPoint end = new GHPoint(0.002, 0.002);
-
-        GHRequest req = new GHRequest().
-                setPoints(Arrays.asList(start, via, end)).
-                setHeadings(Arrays.asList(90., 270., 270.)).
-                setProfile("profile").
-                setPathDetails(Arrays.asList("edge_key"));
-        GHResponse response = instance.route(req);
-        assertFalse(response.hasErrors());
-        assertArrayEquals(new int[]{0, 1, 2, 3, 4}, calcNodes(instance, response.getAll().get(0)));
-    }
-
-    private GraphHopper createSquareGraphInstance() {
-        CarFlagEncoder carEncoder = new CarFlagEncoder();
-        EncodingManager encodingManager = EncodingManager.create(carEncoder);
-        Weighting weighting = new FastestWeighting(carEncoder);
-        GraphHopperStorage g = new GraphBuilder(encodingManager).setCHConfigs(CHConfig.nodeBased("p", weighting)).create();
-
-        //   2---3---4
-        //  /    |    \
-        //  1----8----5
-        //  /    |    /
-        //  0----7---6
-        NodeAccess na = g.getNodeAccess();
-        na.setNode(0, 0.000, 0.000);
-        na.setNode(1, 0.001, 0.000);
-        na.setNode(2, 0.002, 0.000);
-        na.setNode(3, 0.002, 0.001);
-        na.setNode(4, 0.002, 0.002);
-        na.setNode(5, 0.001, 0.002);
-        na.setNode(6, 0.000, 0.002);
-        na.setNode(7, 0.000, 0.001);
-        na.setNode(8, 0.001, 0.001);
-
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(0, 1).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(1, 2).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(2, 3).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(3, 4).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(4, 5).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(5, 6).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(6, 7).setDistance(100));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(7, 0).setDistance(100));
-
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(1, 8).setDistance(110));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(3, 8).setDistance(110));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(5, 8).setDistance(110));
-        GHUtility.setSpeed(60, true, true, carEncoder, g.edge(7, 8).setDistance(110));
-
-        GraphHopper tmp = new GraphHopper() {
-            {
-                setProfiles(new Profile("profile").setVehicle("car").setWeighting("fastest"));
-                loadGraph(g, encodingManager);
-            }
-        };
-        return tmp;
-    }
-
-    @Test
     public void testMultipleCHPreparationsInParallel() {
         HashMap<String, Long> shortcutCountMap = new HashMap<>();
         // try all parallelization modes        
@@ -1174,18 +998,5 @@ public class GraphHopperOSMTest {
         }
     }
 
-    private int[] calcNodes(GraphHopper instance, ResponsePath responsePath) {
-        List<PathDetail> edgeKeys = responsePath.getPathDetails().get("edge_key");
-        int[] result = new int[edgeKeys.size() + 1];
-        for (int i = 0; i < edgeKeys.size(); i++) {
-            int edgeKey = (int) edgeKeys.get(i).getValue();
-            int edgeId = edgeKey / 2;
-            EdgeIteratorState edgeIteratorState = instance.getGraphHopperStorage().getEdgeIteratorState(edgeId, Integer.MIN_VALUE);
-            result[i] = edgeKey % 2 == 0 ? edgeIteratorState.getBaseNode() : edgeIteratorState.getAdjNode();
-            if (i == edgeKeys.size() - 1)
-                result[edgeKeys.size()] = edgeKey % 2 == 0 ? edgeIteratorState.getAdjNode() : edgeIteratorState.getBaseNode();
-        }
-        return result;
-    }
 
 }
